Reorgs rpc directory structure
